movie-raw-data-loader/requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==3.24.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
